requirements.txt
kornia==0.6.2
matplotlib==3.3.4
MedPy==0.4.0
nnunet==1.7.0
numpy==1.19.5
Pillow==9.0.1
runx==0.0.11
scikit_image==0.18.1
skimage==0.0
tensorflow==2.6.0
torch==1.9.0
torchvision==0.10.0
yacs==0.1.8
